% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{deploy_drat}
\alias{deploy_drat}
\title{Deploy drat repo}
\usage{
deploy_drat(drat_repo = paste(git_user(), "/drat", sep = ""),
  commit_message = paste(git_user_repo(),
  ": travis build $TRAVIS_BUILD_NUMBER", sep = ""),
  valid_branches = c("master"), email = "travis@travis-ci.org",
  name = "Travis CI", output_dir = "_drat")
}
\arguments{
\item{drat_repo}{"USER/REPO" location of the drat repository}

\item{commit_message}{string to be used for git commit message}

\item{valid_branches}{string vector of branch names that are allowed to deploy. Defaults to "USER/REPO: travis build $TRAVIS_BUILD_NUMBER"}

\item{email}{email of the 'user' making the travis commit}

\item{name}{name of the 'user' making the travis commit}

\item{output_dir}{location to perform drat functions.  This folder should not exist beforehand}
}
\description{
Publish the package to the drat repo with travis
}
\examples{
\dontrun{
  secure_token(FALSE)
}
}

